+++ src/commands/symlink.js
@@ -1,23 +1,25 @@
 const fs = require('fs')
 const path = require("path")
 const util = require('node:util');
-const exec = util.promisify(require('node:child_process').exec);
+const spawn = require('../spawn');
 
 const cwdPath = path.resolve(process.cwd());
-const cwdNodeModulesPath = path.resolve(cwdPath, 'node_modules')
+let cwdNodeModulesPath = path.resolve(cwdPath, 'node_modules')
 
 
-// let doInstall = process.argv[2];
-// console.log('doInstall', doInstall)
 let reposLength, failed = [];
 
 module.exports = async function symlink(repos) {
     reposLength = repos.length
 
     for (let i = 0; i < repos.length; i++) {
-        if (repos[i].install == true) {
+ 
+        if (cwdPath === repos[i].absolutePath && !fs.existsSync(cwdNodeModulesPath)) {
+            await install(repos[i], repos)    
             reposLength -= 1
-            await install(repos[i])
+        } else if (repos[i].install == true) {
+            reposLength -= 1
+            await install(repos[i], repos)
         } else if (cwdPath !== repos[i].absolutePath) {
             await createSymlink(repos[i])
         }
@@ -76,29 +78,31 @@ function erSymlink(name, dest){
 }
 
 
-async function install(repo) {
+async function install(repo, repos) {
     let dpath = repo.absolutePath
     if (!fs.existsSync(dpath)) {
         failed.push({name: 'install', des: 'path doesn\'t exist:' + dpath})
         return console.error(dpath, 'not exist')
     }
     try {
-        // let node_modules = path.resolve(dpath, 'node_modules');
-        // if (fs.existsSync(node_modules)) {
-        //     fs.rm(node_modules, { recursive: true, force: true }, function (err) {
-        //         if (err) {
-        //             console.log('failed');
-        //         } else {
-        //             erSymlink(repo.name, dest)
-        //         }
-        //     });
-        // }
-        let {error} = await exec(`${repo.packageManager} install `, { cwd: dpath })
-        if (!error)
-            console.log(repo.name, 'installed')
-        else {
-            failed.push({name: 'install ', des: error})
-            console.error(repo.name, 'failed to install', error)
+        console.log('installing', repo.name)
+        let exitCode = await spawn(repo.packageManager, ['install'], {
+            cwd: repo.absolutePath,
+            shell: true,
+            stdio: 'inherit'
+        });
+        
+        if (exitCode !== 0) {
+            failed.push({
+                name: repo.name,
+                des: `${repo.packageManager} install failed`
+            })
+            console.error(`${repo.name}: ${repo.packageManager} install failed`.red)
+        } else {
+            let linkFailed = await require('./link.js')([repo], repos)
+            if (linkFailed)
+                failed.push(linkFailed)
+
         }
 
     }